# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "backend part of test constructor",
        "title": "test constructor",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/admin/manager/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создать нового организатора",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Создать организатора",
                "parameters": [
                    {
                        "description": "Manager object",
                        "name": "test",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.CreateManagerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/admin.CreateManagerResponse"
                        }
                    }
                }
            }
        },
        "/api/intern/attempt/finish": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получение ответов стажёра",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "intern"
                ],
                "summary": "Завершить тест",
                "parameters": [
                    {
                        "description": "Answers object",
                        "name": "answers",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/intern.FinishAttemptRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/intern.FinishAttemptResponse"
                        }
                    }
                }
            }
        },
        "/api/intern/tests/{link}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создание попытки",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "intern"
                ],
                "summary": "Пройти тест",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ссылка теста",
                        "name": "link",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/intern.StartAttemptResponse"
                        }
                    }
                }
            }
        },
        "/api/manager/tests": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Получить список тестов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manager"
                ],
                "summary": "Получить тесты для организотора",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/manager.TestsInfoResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создать новый тест",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manager"
                ],
                "summary": "Создать тест",
                "parameters": [
                    {
                        "description": "Test object",
                        "name": "test",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/manager.CreateTestInfo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/manager/tests/delete/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manager"
                ],
                "summary": "Удалить тест",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID теста",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Логин",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Вход",
                "parameters": [
                    {
                        "description": "User object",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginResponse"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Создать нового пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Создать пользователя",
                "parameters": [
                    {
                        "description": "User object",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/auth.RegisterResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "admin.CreateManagerRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "admin.CreateManagerResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "auth.LoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth.LoginResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "auth.RegisterRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "auth.RegisterResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "intern.FinishAttemptRequest": {
            "type": "object",
            "properties": {
                "userAnswers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/intern.UserAnswerInfo"
                    }
                }
            }
        },
        "intern.FinishAttemptResponse": {
            "type": "object",
            "properties": {
                "max_test_points": {
                    "type": "integer"
                },
                "passed": {
                    "type": "boolean"
                },
                "result": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                }
            }
        },
        "intern.PublicMatching": {
            "type": "object",
            "properties": {
                "left": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "right": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "intern.PublicOptions": {
            "type": "object",
            "properties": {
                "case_sensitive": {
                    "type": "boolean"
                },
                "choice": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "matching": {
                    "$ref": "#/definitions/intern.PublicMatching"
                },
                "sequence": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "intern.QuestionInfo": {
            "type": "object",
            "properties": {
                "options": {
                    "$ref": "#/definitions/intern.PublicOptions"
                },
                "order_number": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.QType"
                }
            }
        },
        "intern.StartAttemptResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/intern.QuestionInfo"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "intern.UserAnswer": {
            "type": "object",
            "properties": {
                "choices": {
                    "type": "array",
                    "items": {
                        "type": "boolean"
                    }
                },
                "matching": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MatchingPair"
                    }
                },
                "sequence": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SequenceItem"
                    }
                },
                "user_input": {
                    "type": "string"
                }
            }
        },
        "intern.UserAnswerInfo": {
            "type": "object",
            "properties": {
                "answer": {
                    "$ref": "#/definitions/intern.UserAnswer"
                },
                "question_id": {
                    "type": "integer"
                }
            }
        },
        "manager.CreateQuestionInfo": {
            "type": "object",
            "properties": {
                "options": {
                    "$ref": "#/definitions/models.QuestionOptions"
                },
                "order_number": {
                    "type": "integer"
                },
                "points": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "manager.CreateTestInfo": {
            "type": "object",
            "properties": {
                "complete_time": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "fail_text": {
                    "type": "string"
                },
                "is_percentage": {
                    "type": "boolean"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/manager.CreateQuestionInfo"
                    }
                },
                "success_text": {
                    "type": "string"
                },
                "threshold": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "manager.TestInfo": {
            "type": "object",
            "properties": {
                "complete_time": {
                    "type": "integer"
                },
                "creator_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "fail_text": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_percentage": {
                    "type": "boolean"
                },
                "success_text": {
                    "type": "string"
                },
                "test_id": {
                    "type": "integer"
                },
                "test_link": {
                    "type": "string"
                },
                "threshold": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "manager.TestsInfoResponse": {
            "type": "object",
            "properties": {
                "tests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/manager.TestInfo"
                    }
                }
            }
        },
        "models.ChoiceOption": {
            "type": "object",
            "properties": {
                "is_true": {
                    "type": "boolean"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "models.MatchingPair": {
            "type": "object",
            "properties": {
                "left": {
                    "type": "string"
                },
                "right": {
                    "type": "string"
                }
            }
        },
        "models.QType": {
            "type": "string",
            "enum": [
                "single_choice",
                "multiple_choice",
                "text_input",
                "matching",
                "correct_order"
            ],
            "x-enum-varnames": [
                "SingleChoice",
                "MultipleChoice",
                "TextInput",
                "Matching",
                "CorrectOrder"
            ]
        },
        "models.QuestionOptions": {
            "type": "object",
            "properties": {
                "case_sensitive": {
                    "type": "boolean"
                },
                "choice": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ChoiceOption"
                    }
                },
                "correct_input": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "matching": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MatchingPair"
                    }
                },
                "sequence": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SequenceItem"
                    }
                }
            }
        },
        "models.SequenceItem": {
            "type": "object",
            "properties": {
                "order": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}